+++ CHANGELOG.md
@@ -12,3 +12,11 @@ this project uses *EVEN* minor version numbers for release versions and *ODD* mi
 ## [0.9] 2024-02-28
 
 - Initial preview release
+
+## Unreleased
+
+### Changed
+
+- DBOS Cloud integration now uses Cloud API directly instead of `dbos-cloud` CLI.
+- `Delete Stored Application Database Passwords` command renamed to `Delete Stored Passwords`
+- `Delete Stored Passwords` command deletes stored DBOS Cloud credentials as well as cloud hosted database passwords.